+++ src/version/index.js
@@ -6,15 +6,15 @@
  * DO NOT EDIT MANUALLY - run 'npm run build' to update.
  *
  * @module version
- * @generated 2026-03-13
+ * @generated 2026-04-09
  * @see scripts/build.js
  */
 
 // Version info is injected at build time
 // DO NOT EDIT - this is auto-generated
-export const VERSION = '15.9.0';
-export const BUILD_DATE = '2026-03-13';
-export const BUILD_TIMESTAMP = 1773437216157;
+export const VERSION = '15.10.0';
+export const BUILD_DATE = '2026-04-09';
+export const BUILD_TIMESTAMP = 1775761837718;
 
 /**
  * Version information object